+++ src/lib.rs
@@ -3,14 +3,14 @@ extern crate lazy_static;
 pub mod auth;
 pub mod cli;
 pub mod config;
-pub mod device_update_import;
+pub mod device_update;
 pub mod file;
 pub mod ssh;
 mod validators;
 use anyhow::{Context, Result};
 use cli::{
     Command,
-    FileConfig::{CopyFromImage, CopyToImage},
+    File::{CopyFromImage, CopyToImage},
     IdentityConfig::{
         SetConfig, SetDeviceCertificate, SetIotLeafSasConfig, SetIotedgeGatewayConfig,
     },
@@ -169,7 +169,50 @@ pub fn run() -> Result<()> {
         }) => run_image_command(image, generate_bmap, compress_image, |img: &PathBuf| {
             file::set_iot_hub_device_update_config(&iot_hub_device_update_config, img)
         })?,
-        Command::IotHubDeviceUpdate(IotHubDeviceUpdate::Import) => device_update_import::import()?,
+        Command::IotHubDeviceUpdate(IotHubDeviceUpdate::ImportUpdate {
+            import_manifest: import_manifest_path,
+            storage_container_name,
+            tenant_id,
+            client_id,
+            client_secret,
+            instance_id,
+            device_update_endpoint_url,
+            blob_storage_account,
+            blob_storage_key,
+        }) => device_update::import_update(
+            &import_manifest_path,
+            storage_container_name,
+            tenant_id,
+            client_id,
+            client_secret,
+            instance_id,
+            &device_update_endpoint_url,
+            blob_storage_account,
+            blob_storage_key,
+        )?,
+        Command::IotHubDeviceUpdate(IotHubDeviceUpdate::CreateImportManifest {
+            image,
+            script,
+            manufacturer,
+            model,
+            compatibilityid,
+            provider,
+            consent_handler,
+            swupdate_handler,
+            distro_name,
+            version,
+        }) => device_update::create_import_manifest(
+            &image,
+            &script,
+            manufacturer,
+            model,
+            compatibilityid,
+            provider,
+            consent_handler,
+            swupdate_handler,
+            distro_name,
+            version,
+        )?,
         Command::Ssh(SetConnection {
             device,
             username,