+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.20.0] Q1 2024
+- device update import command:
+  - splitted import command into create-import-manifest and import-update
+  - added test case for create-import-manifest command
+- documentation: reduced examples in Readme.md but therefore improved comments in clap cli
+
 ## [0.19.3] Q1 2024
 - update default credentials for keycloak backend authentication